UI: Commit info consistent across screens
